Implemented the necessary base functionality to support m3 files with content in different versions.
That the structure.xml file no longer specifies via refTo field
the version, since the same version of a structure can have childs
of different versions (in different files).

The generated xml files now provide the necessary version information,
so that it's possible to select the proper structures when
reconstructing the model from the xml file.